+++ packages/atomic/src/components/facets-v1/atomic-rating-facet/atomic-rating-facet.tsx
@@ -1,4 +1,4 @@
-import {Component, h, State, Prop, VNode, Host} from '@stencil/core';
+import {Component, h, State, Prop, VNode, Host, Element} from '@stencil/core';
 import {
   NumericFacet,
   buildNumericFacet,
@@ -22,10 +22,11 @@ import {FacetContainer} from '../facet-container/facet-container';
 import {FacetHeader} from '../facet-header/facet-header';
 import {FacetValueCheckbox} from '../facet-value-checkbox/facet-value-checkbox';
 import {FacetValueLink} from '../facet-value-link/facet-value-link';
-import {FacetValueIconRating} from '../facet-value-icon-rating/facet-value-icon-rating';
+import {Rating} from '../../atomic-rating/atomic-rating';
 import {BaseFacet} from '../facet-common';
 import Star from '../../../images/star.svg';
 import {Schema, StringValue} from '@coveo/bueno';
+import {registerFacetToStore} from '../../../utils/store';
 
 /**
  * A facet is a list of values for a certain field occurring in the results, ordered using a configurable criteria (e.g., number of occurrences).
@@ -63,6 +64,7 @@ export class AtomicRatingFacet
   @InitializeBindings() public bindings!: Bindings;
   public facet!: NumericFacet;
   public searchStatus!: SearchStatus;
+  @Element() private host!: HTMLElement;
 
   @BindStateToController('facet')
   @State()
@@ -71,7 +73,6 @@ export class AtomicRatingFacet
   @State()
   public searchStatusState!: SearchStatusState;
   @State() public error!: Error;
-  @State() public isCollapsed = false;
 
   /**
    * Specifies a unique identifier for the facet.
@@ -110,6 +111,10 @@ export class AtomicRatingFacet
    * - Use a stringified SVG to display it directly.
    */
   @Prop() public icon = Star;
+  /**
+   * Specifies if the facet is collapsed.
+   */
+  @Prop({reflect: true, mutable: true}) public isCollapsed = false;
 
   private validateProps() {
     new Schema({
@@ -136,10 +141,12 @@ export class AtomicRatingFacet
     };
     this.facet = buildNumericFacet(this.bindings.engine, {options});
     this.facetId = this.facet.state.facetId;
-    this.bindings.store.state.numericFacets[this.facetId] = {
+    registerFacetToStore(this.bindings.store, 'numericFacets', {
       label: this.label,
-      format: (facetValue) => this.formatFacetValue(facetValue),
-    };
+      facetId: this.facetId!,
+      element: this.host,
+      format: (value) => this.formatFacetValue(value),
+    });
   }
 
   private get scaleFactor() {
@@ -198,11 +205,11 @@ export class AtomicRatingFacet
             i18n={this.bindings.i18n}
             onClick={onClick}
           >
-            <FacetValueIconRating
+            <Rating
               numberOfTotalIcons={this.maxValueInIndex}
               numberOfActiveIcons={facetValue.start}
               icon={this.icon}
-            ></FacetValueIconRating>
+            ></Rating>
           </FacetValueCheckbox>
         );
       case 'link':
@@ -214,11 +221,11 @@ export class AtomicRatingFacet
             i18n={this.bindings.i18n}
             onClick={onClick}
           >
-            <FacetValueIconRating
+            <Rating
               numberOfTotalIcons={this.maxValueInIndex}
               numberOfActiveIcons={facetValue.start}
               icon={this.icon}
-            ></FacetValueIconRating>
+            ></Rating>
           </FacetValueLink>
         );
     }