+++ packages/atomic/CHANGELOG.md
@@ -3,6 +3,166 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.66.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.66.1...@coveo/atomic@0.66.2) (2021-08-18)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+## [0.66.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.66.0...@coveo/atomic@0.66.1) (2021-08-18)
+
+
+### Bug Fixes
+
+* **atomic:** made desktop and mobile [@screen](https://github.com/screen) directives mutually exclusive ([#1105](https://github.com/coveo/ui-kit/issues/1105)) ([b026e9b](https://github.com/coveo/ui-kit/commit/b026e9bc5bee0f53b3f215574f34149e3755b72d))
+
+
+
+
+
+# [0.66.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.65.0...@coveo/atomic@0.66.0) (2021-08-17)
+
+
+### Features
+
+* **atomic:** style atomic-sort-dropdown for v1 & add a sort dropdown to the atomic-refine-modal ([#1102](https://github.com/coveo/ui-kit/issues/1102)) ([de7b1d4](https://github.com/coveo/ui-kit/commit/de7b1d4d57e92f92318b5f27e7ad09ead58f0e1a))
+
+
+
+
+
+# [0.65.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.64.0...@coveo/atomic@0.65.0) (2021-08-17)
+
+
+### Features
+
+* **atomic:** added atomic-format-* support to atomic-result-number ([#1096](https://github.com/coveo/ui-kit/issues/1096)) ([b222565](https://github.com/coveo/ui-kit/commit/b22256574fa716b927f5fe29573c1e70c143c659))
+
+
+
+
+
+# [0.64.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.63.1...@coveo/atomic@0.64.0) (2021-08-17)
+
+
+### Features
+
+* **atomic:** add refine toggle button and modal components, rework v1.html layout ([#1079](https://github.com/coveo/ui-kit/issues/1079)) ([23a0ebd](https://github.com/coveo/ui-kit/commit/23a0ebdc280cfaaaf932f83a4846df91e81ce5f9))
+
+
+
+
+
+## [0.63.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.63.0...@coveo/atomic@0.63.1) (2021-08-16)
+
+
+### Bug Fixes
+
+* **atomic:** fix no results page query encoding characters ([#1088](https://github.com/coveo/ui-kit/issues/1088)) ([709605a](https://github.com/coveo/ui-kit/commit/709605a9a1a0edc2345b196df3ba7a56d222261d))
+
+
+
+
+
+# [0.63.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.62.0...@coveo/atomic@0.63.0) (2021-08-16)
+
+
+### Features
+
+* **atomic:** atomic load more results ([#1077](https://github.com/coveo/ui-kit/issues/1077)) ([67b6bbc](https://github.com/coveo/ui-kit/commit/67b6bbcd58a0c9bbf1b42722e5793e2022837dd8))
+
+
+
+
+
+# [0.62.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.61.0...@coveo/atomic@0.62.0) (2021-08-13)
+
+
+### Features
+
+* **atomic:** added atomic-result-rating component ([#1080](https://github.com/coveo/ui-kit/issues/1080)) ([5d3332f](https://github.com/coveo/ui-kit/commit/5d3332fde519b223ad7b9fcfb20cfcb8b1edbf7b))
+* **atomic:** changed appearance of quickview button ([#1087](https://github.com/coveo/ui-kit/issues/1087)) ([94cde86](https://github.com/coveo/ui-kit/commit/94cde86194d6d159ef2fc9bbdb9dc0ef59c4b187))
+* **atomic:** gave opinionated style to visual section ([2b77bf6](https://github.com/coveo/ui-kit/commit/2b77bf6cc4cfca6cdad2186398a55e18cad895d2))
+
+
+
+
+
+# [0.61.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.60.2...@coveo/atomic@0.61.0) (2021-08-13)
+
+
+### Bug Fixes
+
+* **atomic:** made table refresh when sort or filters are changed ([#1082](https://github.com/coveo/ui-kit/issues/1082)) ([469a4b0](https://github.com/coveo/ui-kit/commit/469a4b0f138a313d3fd091dc2ad271502f074ad0))
+
+
+### Features
+
+* **atomic:** limited height of excerpt section ([#1089](https://github.com/coveo/ui-kit/issues/1089)) ([9280dc7](https://github.com/coveo/ui-kit/commit/9280dc7ed92479c179e0080312e4fb9d57bab831))
+
+
+
+
+
+## [0.60.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.60.1...@coveo/atomic@0.60.2) (2021-08-13)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+## [0.60.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.60.0...@coveo/atomic@0.60.1) (2021-08-13)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+# [0.60.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.59.2...@coveo/atomic@0.60.0) (2021-08-12)
+
+
+### Features
+
+* **atomic:** added padding to table result list ([#1083](https://github.com/coveo/ui-kit/issues/1083)) ([7e64b14](https://github.com/coveo/ui-kit/commit/7e64b14ddc3459ad1b0e0030416312e12a7771ca))
+
+
+
+
+
+## [0.59.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.59.1...@coveo/atomic@0.59.2) (2021-08-12)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+## [0.59.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.59.0...@coveo/atomic@0.59.1) (2021-08-11)
+
+
+### Bug Fixes
+
+* **atomic:** fix text overflow on Category Facet ([5f24fd9](https://github.com/coveo/ui-kit/commit/5f24fd968eb0349b92a2d5b7404460435f2fcbbd))
+
+
+
+
+
+# [0.59.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.58.0...@coveo/atomic@0.59.0) (2021-08-11)
+
+
+### Features
+
+* **atomic:** added atomic-result-fields-list component ([#1073](https://github.com/coveo/ui-kit/issues/1073)) ([af4759a](https://github.com/coveo/ui-kit/commit/af4759a565100706150a79324696a39ed1d3a4ed))
+
+
+
+
+
 # [0.58.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@0.57.0...@coveo/atomic@0.58.0) (2021-08-10)